requirements.txt
fastapi
uvicorn[standard]
openai
anthropic
numpy
tqdm
python-dotenv
PyPDF2
requests
tenacity